split the exmaple file into multiple files
